(PC-10877) Remove fullAddress from useOffer and create useFormatFullAddress hook
