+++ src/libs/address/__mocks__/useFormatFullAddress.ts
@@ -1,0 +1,3 @@
+import { mockedFullAddress } from 'libs/address/fixtures/mockedFormatFullAddress'
+
+export const formatFullAddress = () => mockedFullAddress
