+++ src/libs/address/fixtures/mockedFormatFullAddress.ts
@@ -1,0 +1,1 @@
+export const mockedFullAddress = 'PATHE BEAUGRENELLE, 2 RUE LAMENNAIS, 75008 PARIS'
